refactor/ clean code

+++ Assets/Scripts/Controller/CameraController.cs
@@ -5,6 +5,6 @@ public class CameraController : Singleton<CameraController>
 {
     public void SetPosition(Vector2Int mapSize)
     {
-        transform.position = new Vector3((float)(mapSize.x - 1) / 2f, (float)(mapSize.y - 1) / 2, -10f);
+        transform.position = new Vector3((mapSize.x - 1) / 2f, (float)(mapSize.y - 1) / 2, -10f);
     }
 }
